tob-wallet/requirements.txt
# Main
Django==1.11.10
django-allauth==0.35.0
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-haystack==2.7.0
django-rest-auth==0.9.3
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0

# Database managment
psycopg2>=2.7.3.1,<3

drf-haystack==1.6.1

# Web Server
gunicorn==19.7.1
httpie==0.9.9
PyJWT==1.5.3
pytest==3.4.1
pytest-asyncio==0.8.0
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.6.1
simplejson==3.13.2
uritemplate==3.0.0
requests==2.7.0
django-request-logging==0.5.5
